fix: use `abi.encode` for hashing arrays (XMTP-10) (#165)
Change hashing of nodeIds_ to use abi.encode in PayerReportManager to address XMTP-10
This pull request updates the hashing logic in the PayerReportManager contract to use abi.encode for nodeIds_ when computing the keccak256 that combines endSequenceId_, endMinuteSinceEpoch_, payersMerkleRoot_, and the nodeIds_ hash. It also updates unit tests to reflect the new computed values.

Modify PayerReportManager to replace abi.encodePacked(nodeIds_) with abi.encode(nodeIds_) in the hash computation in PayerReportManager.sol
Update three expected bytes32 constants in tests to match the new hash results in PayerReportManager.t.sol